Restructure the data a bit so that we do not have to load every round
