+++ clients/crunchbase/crunchbase.noms
@@ -25,27 +25,3 @@ struct Round {
     FundedAt:               Int64
     RaisedAmountUsd:        Float64
 }
-
-enum QuarterEnum {
-    Q1
-    Q2
-    Q3
-    Q4
-}
-
-struct Quarter {
-    Year:    Int32
-    Quarter: QuarterEnum
-}
-
-struct Key {
-    union {
-        Year:     Int32
-        Quarter:  Quarter
-        Category: String
-        Region:   String
-    }
-}
-
-using Map(String, Ref(Company))
-using Map(Ref(Key), Set(Ref(Round)))
