+++ clients/crunchbase/importer/rungen.go
@@ -1,3 +1,3 @@
 package main
 
-//go:generate go run ../../../nomdl/codegen/codegen.go -package=main -in=../crunchbase.noms -out-dir=.
+//go:generate go run ../../../nomdl/codegen/codegen.go -package=main -out-dir=.
